Cleanup SketchPoint and SketchPass classes; Move to different files (#1147)
